Fix check_ssl_connection() function
Function doesn't wait for `openssl s_client ...` to finish.
It assumes that when the command is still running that is the
successful condition. However the function should wait for
exit code from the binary. We saw in production intermittent
and very often `skale ssl upload` failures. This change
should fix this problem and underlying race condition.